adjusted highscore saving logic to only save new high scores.

+++ src/services/highscore-manager.ts
@@ -10,6 +10,10 @@ const loadHighscores = (): HighScore[] => {
 
 const saveHighScoreByName = (name: string, score: number): void => {
   let highscores = loadHighscores();
+  const existingScore = highscores.find((hscore) => hscore.name === name);
+  if (existingScore && existingScore.score > score) {
+    return;
+  }
   highscores = highscores.filter((hscore) => hscore.name !== name);
   highscores.push({ name, score });
   localStorage.setItem("highscores", JSON.stringify(highscores));
